refactor: upd summary info
