--- src/components/Techs/Techs.js
@@ -1,22 +1,0 @@
-import './Techs.css';
-
-export default function Techs() {
-  return (
-    <section className="techs" id="techs">
-      <h2 className="techs__title">Технологии</h2>
-      <div className="techs__wrap">
-        <h2 className="techs__subtitle">7 технологий</h2>
-        <p className="techs__info">На курсе веб-разработки мы освоили технологии, которые применили в дипломном проекте.</p>
-        <ul className="techs__list">
-          <li className="techs__list-item">HTML</li>
-          <li className="techs__list-item">CSS</li>
-          <li className="techs__list-item">JS</li>
-          <li className="techs__list-item">React</li>
-          <li className="techs__list-item">Git</li>
-          <li className="techs__list-item">Express.js</li>
-          <li className="techs__list-item">mongoDB</li>
-        </ul>
-      </div>
-    </section >
-  );
-}
